+++ client/src/api/block_builder/input_selection/manual.rs
@@ -5,11 +5,11 @@
 
 use std::collections::HashSet;
 
-use bee_block::{
+use crypto::keys::slip10::Chain;
+use iota_types::block::{
     address::Address,
     output::{AliasId, Output, RentStructure},
 };
-use crypto::keys::slip10::Chain;
 
 use crate::{
     api::{