+++ client/src/lib.rs
@@ -54,10 +54,9 @@ pub mod secret;
 pub mod stronghold;
 pub mod utils;
 
-pub use bee_api_types as api_types;
-pub use bee_block as block;
 pub use bee_pow as pow;
 pub use crypto::{self, keys::slip10::Seed};
+pub use iota_types::{api as api_types, block};
 pub use packable;
 pub use url::Url;