+++ client/src/secret/mnemonic.rs
@@ -6,14 +6,14 @@
 use std::ops::Range;
 
 use async_trait::async_trait;
-use bee_block::{
-    address::{Address, Ed25519Address},
-    signature::{Ed25519Signature, Signature},
-    unlock::{SignatureUnlock, Unlock},
-};
 use crypto::{
     hashes::{blake2b::Blake2b256, Digest},
     keys::slip10::{Chain, Curve, Seed},
+};
+use iota_types::block::{
+    address::{Address, Ed25519Address},
+    signature::{Ed25519Signature, Signature},
+    unlock::{SignatureUnlock, Unlock},
 };
 
 use super::{types::InputSigningData, GenerateAddressMetadata, SecretManage};