+++ client/CHANGELOG.md
@@ -19,7 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
-## 2.0.1-rc.2 - YYYY-MM-DD
+## 2.0.1-rc.3 - 2022-XX-XX
+
+### Changed
+
+- Use new `iota-types` crate instead of `bee-api-types` and `bee-block`;
 
 ### Fixed